rgw: move resharding code to BucketReshard class
